chore: fix build bundles
 - remove rns from bundle externals and add redux
 - move NotifyProvider into its own file to remove circular depenency with NotifyContext and NotifyPortal
